Rename ga.ts and ga.test.ts, update imports
