+++ packages/manager/src/components/LinodeCLIModal/LinodeCLIModal.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { CopyTooltip } from 'src/components/CopyTooltip/CopyTooltip';
 import { Dialog } from 'src/components/Dialog/Dialog';
-import { sendCLIClickEvent } from 'src/utilities/ga';
+import { sendCLIClickEvent } from 'src/utilities/analytics';
 import { styled } from '@mui/material/styles';
 
 export interface ImageUploadSuccessDialogProps {
