+++ packages/manager/src/features/Domains/DomainsEmptyLandingPage.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import DomainIcon from 'src/assets/icons/entityIcons/domain.svg';
 import { ResourcesSection } from 'src/components/EmptyLandingPageResources/ResourcesSection';
-import { sendEvent } from 'src/utilities/ga';
+import { sendEvent } from 'src/utilities/analytics';
 import {
   gettingStartedGuides,
   headers,
