+++ packages/manager/src/features/Linodes/LinodeEntityDetail.tsx
@@ -29,7 +29,7 @@ import { useRegionsQuery } from 'src/queries/regions';
 import { useTypeQuery } from 'src/queries/types';
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
 import formatDate from 'src/utilities/formatDate';
-import { sendLinodeActionMenuItemEvent } from 'src/utilities/ga';
+import { sendLinodeActionMenuItemEvent } from 'src/utilities/analytics';
 import { pluralize } from 'src/utilities/pluralize';
 import { ipv4TableID } from './LinodesDetail/LinodeNetworking/LinodeNetworking';
 import { lishLink, sshLink } from './LinodesDetail/utilities';
