+++ packages/manager/src/features/Linodes/LinodesCreate/LinodeCreateContainer.tsx
@@ -56,7 +56,7 @@ import { upsertLinode } from 'src/store/linodes/linodes.actions';
 import { MapState } from 'src/store/types';
 import { getAPIErrorOrDefault } from 'src/utilities/errorUtils';
 import { isEURegion } from 'src/utilities/formatRegion';
-import { sendCreateLinodeEvent, sendEvent } from 'src/utilities/ga';
+import { sendCreateLinodeEvent, sendEvent } from 'src/utilities/analytics';
 import { getQueryParamsFromQueryString } from 'src/utilities/queryParams';
 import scrollErrorIntoView from 'src/utilities/scrollErrorIntoView';
 import { validatePassword } from 'src/utilities/validatePassword';
