+++ packages/manager/src/features/ObjectStorage/BucketDetail/BucketDetail.tsx
@@ -14,7 +14,7 @@ import { getQueryParamFromQueryString } from 'src/utilities/queryParams';
 import { OBJECT_STORAGE_DELIMITER } from 'src/constants';
 import { ObjectDetailsDrawer } from './ObjectDetailsDrawer';
 import { ObjectUploader } from '../ObjectUploader/ObjectUploader';
-import { sendDownloadObjectEvent } from 'src/utilities/ga';
+import { sendDownloadObjectEvent } from 'src/utilities/analytics';
 import { Table } from 'src/components/Table';
 import { TableBody } from 'src/components/TableBody';
 import { TableCell } from 'src/components/TableCell';
