+++ packages/manager/src/features/ObjectStorage/BucketLanding/BucketLandingEmptyState.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { ResourcesSection } from 'src/components/EmptyLandingPageResources/ResourcesSection';
-import { sendEvent } from 'src/utilities/ga';
+import { sendEvent } from 'src/utilities/analytics';
 import { StyledBucketIcon } from './StylesBucketIcon';
 import { useHistory } from 'react-router-dom';
 import {
