+++ packages/manager/src/utilities/emptyStateLandingUtils.ts
@@ -1,4 +1,4 @@
-import { sendEvent } from 'src/utilities/ga';
+import { sendEvent } from 'src/utilities/analytics';
 
 export const guidesMoreLinkText = 'Check out all our Docs';
 export const docsLink = 'https://www.linode.com/docs/';
